Publisher/requirements.txt
azure-iot-device==2.12.0
python-dotenv
avro-python3